RYSK-349: Extract to util.
