+++ packages/front-end/src/components/VaultPerformance/VaultPerformance.tsx
@@ -4,7 +4,6 @@ import { gql, useQuery } from "@apollo/client";
 import { AnimatePresence } from "framer-motion";
 import { useEffect, useState } from "react";
 
-import { toTwoDecimalPlaces } from "src/utils/rounding";
 
 import { QueriesEnum } from "src/clients/Apollo/Queries";
 import { logError } from "src/utils/logError";
@@ -14,6 +13,7 @@ import { Error } from "./subcomponents/Error";
 import { FadeWrapper } from "./subcomponents/FadeWrapper";
 import { Loading } from "./subcomponents/Loading";
 import { Stats } from "./subcomponents/Stats";
+import { parseData } from "./utils/parseData";
 
 export const VaultPerformance = () => {
   const [chartData, setChartData] = useState<ChartData[]>([]);
@@ -39,74 +39,9 @@ export const VaultPerformance = () => {
   );
 
   useEffect(() => {
-    if (data) {
-      const { pricePerShares } = data;
-      const lastIndex = pricePerShares[pricePerShares.length - 1];
-
-      const pricePerSharesWithPrediction = [
-        ...pricePerShares,
-        {
-          epoch: (parseFloat(lastIndex.epoch) + 1).toString(),
-          growthSinceFirstEpoch: "",
-          predictedGrowthSinceFirstEpoch: "-2.291", // Use lens value.
-          timestamp: (
-            parseFloat(lastIndex.timestamp) + SECONDS_IN_SEVEN_DAYS
-          ).toString(),
-          __typename: "",
-        },
-      ];
-
-      const publicLaunchOffset = pricePerSharesWithPrediction.length
-        ? parseFloat(pricePerSharesWithPrediction[0].growthSinceFirstEpoch)
-        : 0;
-
-      // Values need replacing with API/Chain data.
-      const ethPrices = [
-        1892.21, 1877.3, 1845.48, 1842.73, 1653.45, 1647.6, 1633.62,
-      ];
-
-      const adjustedChartData = pricePerSharesWithPrediction.map(
-        (pricePoint, index, array) => {
-          const pricePointGrowth = parseFloat(pricePoint.growthSinceFirstEpoch);
-          const growthSinceFirstEpoch = toTwoDecimalPlaces(
-            pricePointGrowth - publicLaunchOffset
-          );
-
-          if (pricePoint.predictedGrowthSinceFirstEpoch) {
-            const predictedPricePointGrowth = parseFloat(
-              pricePoint.predictedGrowthSinceFirstEpoch
-            );
-
-            return {
-              ...pricePoint,
-              ethPrice: toTwoDecimalPlaces(
-                (ethPrices[index] / ethPrices[0] - 1) * 100
-              ),
-              growthSinceFirstEpoch: NaN,
-              predictedGrowthSinceFirstEpoch: toTwoDecimalPlaces(
-                predictedPricePointGrowth - publicLaunchOffset
-              ),
-            };
-          }
-
-          if (index === array.length - 2) {
-            return {
-              ...pricePoint,
-              growthSinceFirstEpoch,
-              predictedGrowthSinceFirstEpoch: growthSinceFirstEpoch,
-            };
-          }
-
-          return {
-            ...pricePoint,
-            growthSinceFirstEpoch,
-            predictedGrowthSinceFirstEpoch: null,
-          };
-        }
-      );
-
-      setChartData(adjustedChartData);
-    }
+    parseData(data).then((parsedData) => {
+      if (parsedData) setChartData(parsedData);
+    });
   }, [data]);
 
   return (
